Add a script to run the BDD feature tests
```npm run features```
Calls `chimp` and runs the Gherkin feature tests.
Running the script opens a Firefox browser and tries to connect to a local server.
This will fail as it hasn't been started (or even created) yet.
The three steps for the Gherkin feature show the three possible states:
 - the first step has no content (throws no errors) so this is marked as `passed`
 - the second step tries to connect to a server that's not running (so causes an error) and is marked as `failedd`
 - the third step is `skipped` because an earlier step already failed