Add a basic http server and update the BDD step
The `server.js` file contains code for a very basic HTTP server, which uses the `express` web application framework to respond to GET requests.
The server is set to listen on the local port `3000`.
The first step definition for the BDD test has been changed, and shows another way to start the server (simply `require` the JavaScript file in the test code).

Running the feature test again (using `npm run features`) now fails at the first step with the following error:
    Error: Cannot find module 'express'

In the stack trace we can see a line that looks like the following:
    at Object.<anonymous> (/home/user/media-uploader/server.js:4:17)

This mean the code in `server.js` on line 4 failed. Specifically it couldn't load the express module.
Unlike the `http` module on line 3 above (which is part of NodeJS), the `express` module is external and must first be installed.