Add express.js web application framework module
```npm install express --save```
Similar to when the `chimp` test module was installed, use the NPM command above will install the Express module.
The `--save` switch at the end tells NPM to add the module as a standard dependency in the `package.json` file (note the new `dependencies` object).

Running the tests again now shows the first two steps passing, and just the last step failing with the following error:
    AssertionError [ERR_ASSERTION]: 'to do' == 'Hello World!'